requirements.txt
# ============================================
# CORE DEPENDENCIES
# ============================================
apache-airflow==2.8.1
apache-airflow-providers-google==10.14.0

# ============================================
# DATA PROCESSING
# ============================================
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2

# ============================================
# API CLIENTS
# ============================================
requests==2.31.0
ccxt==4.2.25

# ============================================
# DATABASE
# ============================================
SQLAlchemy==1.4.51
psycopg2-binary==2.9.9

# ============================================
# DATA QUALITY
# ============================================
great-expectations==0.18.8

# ============================================
# DBT
# ============================================
dbt-core==1.7.7
dbt-bigquery==1.7.4

# ============================================
# MACHINE LEARNING
# ============================================
scikit-learn==1.4.0
xgboost==2.0.3
mlflow==2.10.0
optuna==3.5.0
shap==0.44.1

# ============================================
# GOOGLE CLOUD (versões compatíveis)
# ============================================
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
google-cloud-aiplatform==1.40.0
google-auth>=2.23.3  # CORRIGIDO: >=2.23.3 (exigido por google-cloud-storage)

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
pyyaml==6.0.1
python-telegram-bot==20.7

# ============================================
# DEVELOPMENT
# ============================================
jupyter==1.0.0
notebook==7.0.7
matplotlib==3.8.2
plotly==5.18.0
